backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
qdrant-client==1.8.0
pydantic-settings==2.2.1
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
python-multipart==0.0.6
python-dotenv
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
httpx
